Switches to download ZeroMQ dependency in CI
With the ZeroMQ library as downloaded, the build process failed with
some errors that suggest that the ZeroMQ version of the package is
incompatible with FLESnet. Hence, I switched to use the option of
FLESnet's build process via cmake to download, build and use the correct
version via git instead of using ZeroMQ as installed on the maschine.