Adds actual FLESnet build commands to CI
Updating and upgrading ubuntu, and installing FLESnet dependencies was
moved to the before_script as these steps are not part of the actual
build process, but rather a preparation of the environment in which the
build process takes place. In order to make this visible in the log
files of the job execution, explicitly calling these commands in the
before_script is best practice.