Fixes doxygen warnings related to glossary
Apparently, it was once attempted to build a glossary with an approach
resembling an answer to the following stackoverflow question:
https://stackoverflow.com/questions/37410255/is-it-possible-to-create-a-central-glossary-with-doxygen

In order to silence the warnings related to that broken approach an
alias is defined to effectively make doxygen ignore that command. Note,
that the '\' character needs to be escaped as otherwise cmake will not
properly pass the desired value to doxygen and throw an error instead.